+++ app/api/routes/search_routes.py
@@ -21,6 +21,7 @@ from app.crawler.spider_lair.spiders import caerostris_darwini, theraphosidae
 import re
 import json
 from datetime import datetime
+from itertools import permutations
 from flask import Blueprint, request
 from app.models import db, History
 from app.forms import SearchForm
@@ -98,7 +99,7 @@ def scrape_with_crochet(raw_query):
 
     In regard to the partitioned query's regular expression:
         - Assert string is not immediately preceded by any word characters (negative lookbehind): `(?<!\w)`.
-        - Only match match strings followed by specified characters: `[\s|.|,|?|!|:|;|\u2010|\u2013|\u2014]` (i.e., a
+        - Only match strings followed by specified characters: `[\s|.|,|?|!|:|;|\u2010|\u2013|\u2014]` (i.e., a
           space or punctuation).
         - Unicode: \u2010 (hyphen), \u2013 (en-dash), \u2014 (em-dash).
     """
@@ -106,16 +107,17 @@ def scrape_with_crochet(raw_query):
     trunc_amt_2 = 16
 
     raw_query_str_list = raw_query.split()
+    broad_crawler_str_list = []
+    query_permutations = []
     broad_crawler_str = ''
-    deep_crawler_str = ''
-    deep_crawler_str_list = []
 
     for i in range(len(raw_query_str_list)):
         raw_query_substring = raw_query_str_list[i]
-        if raw_query_substring not in stop_word_set:
-            broad_crawler_str += f'|(?<!\w){raw_query_substring}[\s|.|,|?|!|:|;|\u2010|\u2013|\u2014]'
-            deep_crawler_str += f'{raw_query_substring} ' if i < (len(raw_query_str_list) - 1) else raw_query_substring
-            deep_crawler_str_list.append(raw_query_substring)
+        broad_crawler_str_list.append(raw_query_substring)
+        if i <= 4: query_permutations += [' '.join(perm) for perm in permutations(raw_query_str_list, i + 1)]
+        if raw_query_substring not in stop_word_set: broad_crawler_str += f'(?<!\w){raw_query_substring}[\s|.|,|?|!|:|;|\u2010|\u2013|\u2014]'
+        if i < len(raw_query_str_list) - 1: broad_crawler_str += '|'
+
 
     broad_crawler_query_regex = re.compile(rf'{broad_crawler_str}', re.I)
     dispatcher.connect(_crawler_result, signal=signals.item_scraped)
@@ -124,7 +126,7 @@ def scrape_with_crochet(raw_query):
 
     if len(broad_crawler_str):
         for broad_crawler in broad_crawlers: crawl_runner.crawl(broad_crawler, query_regex=broad_crawler_query_regex, trunc_amt_1=trunc_amt_1)
-        for deep_crawler in deep_crawlers: crawl_runner.crawl(deep_crawler, query_string=deep_crawler_str, query_list=deep_crawler_str_list, trunc_amt_1=trunc_amt_1, trunc_amt_2=trunc_amt_2)
+        for deep_crawler in deep_crawlers: crawl_runner.crawl(deep_crawler, query_list=broad_crawler_str_list, query_perms=query_permutations, trunc_amt_1=trunc_amt_1, trunc_amt_2=trunc_amt_2)
         eventual = crawl_runner.join()
         return eventual