+++ react-app/src/components/Search_Page/Search_Results_Page.js
@@ -14,8 +14,8 @@ const SearchResultsPage = ({ style }) => {
   if (user && !user.errors) isUser = true;
   else isUser = false;
 
-  const visitHandler = async (e) => {
-    await dispatch(createVisitEntry({
+  const visitHandler = (e) => {
+    dispatch(createVisitEntry({
       visit: e.target.href,
       updatedAt: new Date().toString(),
       user,
@@ -82,8 +82,10 @@ const SearchResultsPage = ({ style }) => {
   };
 
   return (
-      <div className='search-results-page-container' style={{ backgroundColor: style.background_color }}>
-          {results.length ? processFullResults(results, 'save') : processEmptyResults()}
+      <div className='search-results-page-container'>
+          <div className='search-results-container' style={{ backgroundColor: style.background_color }}>
+              {results.length ? processFullResults(results, 'save') : processEmptyResults()}
+          </div>
       </div>
   );
 };