+++ react-app/src/components/Settings_Page/Edit_Theme_Form.js
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { editProfile } from '../../store/session';
@@ -10,8 +10,11 @@ const EditThemeForm = ({ style }) => {
   const user = useSelector((state) => state.session.user);
   const settingsObj = useSelector((state) => state.settings);
 
-  const [backgroundMedia, setBackgroundMedia] = useState(style.background_media);
-  const [backgroundMediaLoading, setBackgroundMediaLoading] = useState(false);
+  const [render, rerender] = useState(false);
+  const [unlockedThemes, setUnlockedThemes] = useState(new Set());
+  const [clicked, setClicked] = useState(false);
+
+  useEffect(() => {}, [render]);
 
   const fontFamiliesRaw = dropdownData.fonts;
   const fontFamilies = fontFamiliesRaw.map((fontFamily) => (
@@ -23,125 +26,38 @@ const EditThemeForm = ({ style }) => {
       <option key={fontSize}>{fontSize}</option>
   ));
 
-  const resetHandler = (targID) => {
-    const targForm = document.getElementById(targID);
-    const prev = settingsObj[targID];
-    const targFieldset = targForm.children[0];
-    const targFieldsetKids = Array.from(targForm.children[0].children);
-
-    const lockOpenIcon = document.getElementById(`lock-open-${targID}`);
-    const lockIcon = document.getElementById(`lock-${targID}`);
-    const lockBtn = document.getElementById(`lock-btn-${targID}`);
-    const cancelBtn = document.getElementById(`cancel-btn-${targID}`);
-
-    lockOpenIcon.dataset.visibility = 'false';
-    lockIcon.dataset.visibility = 'true';
-    lockBtn.dataset.locked = 'true';
-    cancelBtn.classList.add('invisible');
-
-    targFieldsetKids.forEach((targKid) => {
-      if (targFieldset) targFieldset.disabled = true;
-      if (targKid.dataset.type === 'bg-media-editor-div') {
-        const mediaInput = targKid.children[1];
-        mediaInput.value = '';
-      }
-      switch (targKid.name) {
-        case 'Theme Name': targKid.value = prev.theme_name; break;
-        case 'Background Color': targKid.value = prev.background_color; break;
-        case 'Background Media': targKid.value = ''; break;
-        case 'Background Rotate': targKid.checked = prev.background_rotate; break;
-        case 'Font Color': targKid.value = prev.font_color; break;
-        case 'Font Family': {
-          const targText = prev.font_family.replace(/,\s/, ' | ');
-          const targKids = targKid.children;
-          const targFamily = Array.from(targKids).find((opt) => opt.text === targText);
-          targFamily.selected = true;
-          break;
-        }
-        case 'Font Size': {
-          targKid.value = prev.font_size;
-          const targNum = prev.font_size.replace('px', '');
-          const targKids = targKid.children;
-          const targSize = Array.from(targKids).find((opt) => opt.text === targNum);
-          targSize.selected = true;
-          break;
-        }
-        case 'Accent 1': targKid.value = prev.accent_1; break;
-        case 'Accent 2': targKid.value = prev.accent_2; break;
-        case 'Accent 3': targKid.value = prev.accent_3; break;
-        default: break;
-      }
-    });
+  const resetHandler = (settingID) => {
+    unlockedThemes.delete(settingID);
+    setUnlockedThemes(unlockedThemes);
+    rerender((prv) => !prv);
   };
 
-  const setBackgroundMediaHandler = (e) => {
-    const file = e.target.files[0];
-    if (file) setBackgroundMedia(file);
-  };
-
-  const editFormHandler = async (e) => {
+  const editFormHandler = (e) => {
     e.preventDefault();
     const targForm = e.target;
-    const targFieldset = targForm.children[0];
-    const targFieldsetKids = Array.from(targForm.children[0].children);
-    const settingID = targForm.id;
-    const lockBtn = document.getElementById(`lock-btn-${settingID}`);
-    const formData = new FormData();
+    const settingID = +targForm.id;
+    const formData = new FormData(e.target);
 
-    if (lockBtn.dataset.locked === 'true') {
-      if (targFieldset) targFieldset.disabled = false;
+    if (unlockedThemes.has(settingID)) {
+      unlockedThemes.delete(settingID);
+      setUnlockedThemes(unlockedThemes);
 
-      const lockIcon = lockBtn.children[0];
-      lockIcon.dataset.visibility = 'false';
+      const backgroundRotate = formData.get('backgroundRotate') === 'on';
+      const fontFamily = formData.get('fontFamily').replace(/\s\|\s/, ', ');
+      const fontSize = `${formData.get('fontSize')}px`;
 
-      const lockOpenIcon = document.getElementById(`lock-open-${settingID}`);
-      lockOpenIcon.dataset.visibility = 'true';
-      lockBtn.dataset.locked = 'false';
+      formData.set('backgroundRotate', backgroundRotate);
+      formData.set('fontFamily', fontFamily);
+      formData.set('fontSize', fontSize);
+      formData.set('settingID', settingID);
+      formData.set('userID', user.id);
 
-      const cancelBtn = document.getElementById(`cancel-btn-${settingID}`);
-      cancelBtn.classList.remove('invisible');
-    } else if (lockBtn.dataset.locked === 'false') {
-      targFieldsetKids.forEach((targKid) => {
-        if (targFieldset) targFieldset.disabled = true;
-
-        const lockOpenIcon = lockBtn.children[1];
-        lockOpenIcon.dataset.visibility = 'false';
-
-        const lockIcon = document.getElementById(`lock-${settingID}`);
-        lockIcon.dataset.visibility = 'true';
-        lockBtn.dataset.locked = 'true';
-
-        const cancelBtn = document.getElementById(`cancel-btn-${settingID}`);
-        cancelBtn.classList.add('invisible');
-
-        if (targKid.dataset.type === 'bg-media-editor-div') {
-          const mediaInput = targKid.children[1];
-          formData.append('backgroundMedia', backgroundMedia);
-          setBackgroundMediaLoading(true);
-          mediaInput.value = '';
-        }
-
-        formData.append('settingID', settingID);
-        formData.append('userID', user.id);
-
-        if (targKid.tagName !== 'BUTTON') {
-          switch (targKid.name) {
-            case 'Theme Name': formData.append('themeName', targKid.value); break;
-            case 'Background Color': formData.append('backgroundColor', targKid.value); break;
-            case 'Background Rotate': formData.append('backgroundRotate', targKid.checked); break;
-            case 'Font Color': formData.append('fontColor', targKid.value); break;
-            case 'Font Family': formData.append('fontFamily', targKid.value.replace(/\s\|\s/, ', ')); break;
-            case 'Font Size': formData.append('fontSize', `${targKid.value}px`); break;
-            case 'Accent 1': formData.append('accent1', targKid.value); break;
-            case 'Accent 2': formData.append('accent2', targKid.value); break;
-            case 'Accent 3': formData.append('accent3', targKid.value); break;
-            default: break;
-          }
-        }
-      });
       dispatch(updateUserSetting(settingID, formData));
-      setBackgroundMediaLoading(false);
+    } else {
+      unlockedThemes.add(settingID);
+      setUnlockedThemes(unlockedThemes);
     }
+    rerender((prv) => !prv);
   };
 
   const copyThemeData = (settingID) => {
@@ -200,7 +116,7 @@ const EditThemeForm = ({ style }) => {
         >
           <fieldset
               className='ef1-fieldset'
-              disabled
+              disabled={!unlockedThemes.has(setting.id)}
               style={{
                 backgroundImage: `url(${setting.background_media})`,
                 backgroundColor: setting.background_color,
@@ -220,19 +136,23 @@ const EditThemeForm = ({ style }) => {
                         color: setting.font_color,
                       }}
                     />
-                  <FontAwesomeIcon
-                      id={`cancel-btn-${setting.id}`}
-                      data-setting-id={`${setting.id}`}
-                      className='invisible'
-                      alt='Cancel Changes'
-                      type='Cancel Changes'
-                      icon='window-close'
+                  <button
+                      type='reset'
+                      data-visibility={unlockedThemes.has(setting.id)}
                       onClick={() => resetHandler(setting.id)}
-                      style={{
-                        backgroundColor: setting.background_color,
-                        color: setting.font_color,
-                      }}
-                    />
+                      style={{ fontSize: style.font_size }}
+                    >
+                      <FontAwesomeIcon
+                          id={`cancel-btn-${setting.id}`}
+                          alt='Cancel Changes'
+                          type='Cancel Changes'
+                          icon='window-close'
+                          style={{
+                            backgroundColor: setting.background_color,
+                            color: setting.font_color,
+                          }}
+                        />
+                  </button>
                   <FontAwesomeIcon
                       alt='Copy Theme Data'
                       title='Copy Theme Data'
@@ -243,52 +163,26 @@ const EditThemeForm = ({ style }) => {
                         color: setting.font_color,
                       }}
                     />
-                  {user.custom_theme !== setting.id
-                        && <FontAwesomeIcon
-                            alt='Unselected Theme'
-                            title='Unselected Theme'
-                            icon='circle'
-                            onClick={() => updateActiveTheme(setting.id, 'custom_theme')}
-                            style={{
-                              backgroundColor: setting.background_color,
-                              color: setting.font_color,
-                            }}
-                        />
-                    }
-                  {user.custom_theme === setting.id
-                        && <FontAwesomeIcon
-                            alt='Selected Theme'
-                            title='Selected Theme'
-                            icon='check-circle'
-                            onClick={() => updateActiveTheme(setting.id, 'custom_theme')}
-                            style={{
-                              backgroundColor: setting.background_color,
-                              color: setting.font_color,
-                            }}
-                        />
-                    }
+                  <FontAwesomeIcon
+                      alt={user.custom_theme === setting.id ? 'Unselect Theme' : 'Select Theme'}
+                      title={user.custom_theme === setting.id ? 'Unselect Theme' : 'Select Theme'}
+                      icon={user.custom_theme === setting.id ? 'check-circle' : 'circle'}
+                      onClick={() => updateActiveTheme(setting.id, 'custom_theme')}
+                      style={{
+                        backgroundColor: setting.background_color,
+                        color: setting.font_color,
+                      }}
+                    />
                   <button
                       id={`lock-btn-${setting.id}`}
-                      data-setting-id={`${setting.id}`}
-                      data-locked='true'
+                      data-locked={!unlockedThemes.has(setting.id)}
+                      style={{ fontSize: style.font_size }}
                     >
                       <FontAwesomeIcon
-                          id={`lock-${setting.id}`}
-                          data-visibility='true'
-                          alt='Unlock Theme'
-                          title='Unlock Theme'
-                          icon='lock'
-                          style={{
-                            backgroundColor: setting.background_color,
-                            color: setting.font_color,
-                          }}
-                        />
-                      <FontAwesomeIcon
-                          id={`lock-open-${setting.id}`}
-                          data-visibility='false'
-                          alt='Lock Theme'
-                          title='Lock Theme'
-                          icon='lock-open'
+                          id={unlockedThemes.has(setting.id) ? `lock-open-${setting.id}` : `lock-${setting.id}`}
+                          alt={unlockedThemes.has(setting.id) ? 'Lock Theme' : 'Unlock Theme'}
+                          title={unlockedThemes.has(setting.id) ? 'Lock Theme' : 'Unlock Theme'}
+                          icon={unlockedThemes.has(setting.id) ? 'lock-open' : 'lock'}
                           style={{
                             backgroundColor: setting.background_color,
                             color: setting.font_color,
@@ -300,7 +194,7 @@ const EditThemeForm = ({ style }) => {
               <input
                   id={`theme-name-editor-${idx}`}
                   type='text'
-                  name='Theme Name'
+                  name='themeName'
                   maxLength='50'
                   placeholder='50 Characters Max'
                   aria-placeholder='50 Characters Max'
@@ -314,7 +208,7 @@ const EditThemeForm = ({ style }) => {
               <label htmlFor={`font-size-editor-${idx}`} style={{ fontSize: setting.font_size }}>Font Size</label>
               <select
                   id={`font-size-editor-${idx}`}
-                  name='Font Size'
+                  name='fontSize'
                   defaultValue={setting.font_size?.replace('px', '')}
                   style={{
                     backgroundColor: setting.background_color,
@@ -327,7 +221,7 @@ const EditThemeForm = ({ style }) => {
               <label htmlFor={`font-family-editor-${idx}`}>Font Family</label>
               <select
                   id={`font-family-editor-${idx}`}
-                  name='Font Family'
+                  name='fontFamily'
                   defaultValue={setting.font_family?.replace(/,\s/, ' | ')}
                   style={{
                     backgroundColor: setting.background_color,
@@ -339,34 +233,37 @@ const EditThemeForm = ({ style }) => {
               </select>
 
               <label htmlFor={`font-color-editor-${idx}`}>Font Color</label>
-              <input id={`font-color-editor-${idx}`} name='Font Color' type='color' defaultValue={setting.font_color} />
+              <input id={`font-color-editor-${idx}`} name='fontColor' type='color' defaultValue={setting.font_color} />
 
               <label htmlFor={`bg-color-editor-${idx}`}>Background Color</label>
-              <input id={`bg-color-editor-${idx}`} name='Background Color' type='color' defaultValue={setting.background_color} />
-
-              <div data-type='bg-media-editor-div'>
-                  <label htmlFor={`bg-media-editor-${idx}`}>{backgroundMedia !== '' ? 'Background Media' : 'Added'}</label>
+              <input id={`bg-color-editor-${idx}`} name='backgroundColor' type='color' defaultValue={setting.background_color} />
+              <div className='background-media-div'>
+                  <label htmlFor={`bg-media-editor-${idx}`}>{setting.background_media ? 'Added' : 'Background Media'}</label>
                   <input
                       id={`bg-media-editor-${idx}`}
-                      name='Background Media'
+                      name='backgroundMedia'
                       type='file'
                       accept='image/png, image/jpg, image/jpeg, image/gif'
-                      onChange={setBackgroundMediaHandler}
                     />
-                  {backgroundMediaLoading && (<span>Loading...</span>)}
               </div>
-
               <label htmlFor={`bg-rotate-editor-${idx}`}>Background Rotate</label>
-              <input id={`bg-rotate-editor-${idx}`} name='Background Rotate' type='checkbox' defaultChecked={setting.background_rotate} />
-
+              <input
+                  id={`bg-rotate-editor-${idx}`}
+                  name='backgroundRotate'
+                  type='checkbox'
+                  defaultChecked={setting.background_rotate}
+                  onPointerDown={() => setClicked(setting.id)}
+                  onPointerUp={() => setClicked(false)}
+                  style={{ cursor: setting.id === clicked ? 'not-allowed' : '' }}
+              />
               <label htmlFor={`accent-1-color-editor-${idx}`} style={{ color: setting.accent_1 }}>Accent 1</label>
-              <input id={`accent-1-color-editor-${idx}`} name='Accent 1' type='color' defaultValue={setting.accent_1} />
+              <input id={`accent-1-color-editor-${idx}`} name='accent1' type='color' defaultValue={setting.accent_1} />
 
               <label htmlFor={`accent-2-color-editor-${idx}`} style={{ color: setting.accent_2 }}>Accent 2</label>
-              <input id={`accent-2-color-editor-${idx}`} name='Accent 2' type='color' defaultValue={setting.accent_2} />
+              <input id={`accent-2-color-editor-${idx}`} name='accent2' type='color' defaultValue={setting.accent_2} />
 
               <label htmlFor={`accent-3-color-editor-${idx}`} style={{ color: setting.accent_3 }}>Accent 3</label>
-              <input id={`accent-3-color-editor-${idx}`} name='Accent 3' type='color' defaultValue={setting.accent_3} />
+              <input id={`accent-3-color-editor-${idx}`} name='accent3' type='color' defaultValue={setting.accent_3} />
           </fieldset>
       </form>
   ));